게시물에 대한 Controller, Service, DAO, DB 로직 연결 완료. 게시물 삭제 시, delStatus와 delDate 조정 및 게시물 조회 시 delStatus에 대한 조건 추가

+++ src/main/java/com/kjh/service/ArticleService.java
@@ -34,12 +34,29 @@ public class ArticleService {
 		
 	}
 
-	public void delete(int id) {
+	public ResultData delete(int id) {
+		
+		Article article = articleDao.getArticleById(id);
+		if(article == null) {
+			return new ResultData("F-1", id + "번 게시물이 존재하지 않습니다.");
+		}
+		
 		articleDao.deleteArticle(id);
+		return new ResultData("S-1", id + "번 게시물이 삭제되었습니다.");
+		
 	}
 
 	public ResultData modify(int id, String title, String body) {
-		Article article = getArticleById("")
+		
+		Article article = articleDao.getArticleById(id);
+		if(article == null) {
+			return new ResultData("F-1", id + "번 게시물이 존재하지 않습니다.");
+		}
+		
+		articleDao.modifyArticle(id, title, body);
+		
+		return new ResultData("S-1", id + "번 게시물이 수정되었습니다.");
+		
 	}
